Adicionando ação ao botão de buscar para realizar os filtros que o usuário inserir

+++ src/main/java/br/com/projetas/crud/car/repository/impl/CarRepositoryImpl.java
@@ -26,7 +26,7 @@ public class CarRepositoryImpl implements CarRepositoryCustom {
     @PersistenceContext
     private EntityManager em;
 
-    public List<Car> findByFilter(String brand, String model, String color, Long maxYear, Long minYear, Float maxPrice, Float minPrice, Boolean isNew) {
+    public List<Car> findByFilter(String brand, String model, String color, Integer maxYear, Integer minYear, Float maxPrice, Float minPrice, Boolean isNew) {
 
         Session session = em.unwrap(Session.class);
